Locale-independent login success detection
State-of-work - runs - start/stop/postprog works

As I currently use a different locale as on my other computer, I coudn't log in to my securbox as it detects the success by a string contained in the response, which is dependent on the locale, if you have en_* it's Englisch and if de_* it's German.
Now the success is detectec by the existance of the link, which you could use to access your files from the web browser if you use it that way.

Fixes
 - Locale dependent login success detection